containers/whisperx/requirements.txt
# PyTorch 패키지 (CPU 버전)
--find-links https://download.pytorch.org/whl/cpu
torch==2.0.0
torchaudio==2.0.1

# WhisperX 및 의존성
whisperx==3.3.1
pydub==0.25.1
ffmpeg-python==0.2.0

# 웹 서버 및 API
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.8

# GCP 통합
google-cloud-storage==2.9.0
google-cloud-pubsub==2.18.0

# 유틸리티
python-multipart==0.0.6
python-dotenv==1.0.0
tenacity==8.2.2

